Enhance UI components with semantic spacing and role-based styles
- Introduced new semantic spacing tokens in `app.css` for consistent layout across components.
- Updated various Blade views to utilize role-based spacing classes for improved design consistency.
- Refactored `StatusCommand` to ensure proper return type casting.
- Removed redundant code in the `admin1` and `countries` Livewire components for better performance and readability.

+++ app/Base/Database/Console/Commands/StatusCommand.php
@@ -28,7 +28,7 @@ class StatusCommand extends IlluminateStatusCommand
     {
         $this->loadAllModuleMigrations();
 
-        return parent::handle();
+        return (int) parent::handle();
     }
 
     /**
